Optimise server and controller methods

+++ backend/controllers/product.controller.js
@@ -57,6 +57,12 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Product not found" });
+  }
   try {
     await Product.findByIdAndDelete(id);
     res
@@ -64,6 +70,6 @@ export const deleteProduct = async (req, res) => {
       .json({ success: true, message: "Product deleted successfully" });
   } catch (error) {
     console.log("Error deleting product", error.message);
-    res.status(404).json({ success: false, message: "Product not found" });
+    res.status(500).json({ success: false, message: "Server error" });
   }
 };
